// frontend/components/api/apiSchemas.ts
/**
 * Generated by @openapi-codegen
 *
 * @version 0.1.0
 */
export type Edge = {
  source: string;
  target: string;
};

export type Graph = {
  directed: boolean;
  multigraph: boolean;
  graph?: void;
  nodes: Node[];
  edges: Edge[];
};

export type HTTPValidationError = {
  detail?: ValidationError[];
};

export type Metric = {
  node_id: string;
  messages_in?: number;
  messages_out?: number;
  consumer_lag?: number;
  consumer_read_rate?: number;
  topic_size?: number;
  replicas?: number;
  replicas_available?: number;
  connector_tasks?: number;
};

export type Node = {
  id: string;
  label: string;
  node_type: string;
  x?: number;
  y?: number;
};

export type NodeInfoListItem = {
  name: string;
  value: string | Record<string, any>;
  type: NodeInfoType;
};

/**
 * An enumeration.
 */
export type NodeInfoType = "json" | "basic" | "link";

export type NodeInformation = {
  node_id: string;
  node_type: NodeTypesEnum;
  info: NodeInfoListItem[];
};

/**
 * An enumeration.
 */
export type NodeTypesEnum =
  | "streaming-app"
  | "connector"
  | "topic"
  | "error-topic"
  | "sink/source";

export type Pipelines = {
  pipelines: string[];
};

export type ValidationError = {
  loc: (string | number)[];
  msg: string;
  type: string;
};